# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Meta-Wall",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "meta-wall.onrender.com",
  "port": "{{PORT}}",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/posts/": {
      "get": {
        "tags": [
          "Posts - 搜尋貼文"
        ],
        "description": "搜尋貼文 API",
        "parameters": [
          {
            "name": "timeSort",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Keyword",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "搜尋成功",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": true
                  },
                  "message": {
                    "type": "string",
                    "example": "搜尋成功"
                  },
                  "post": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "662ec324ad9afd608b5c6d88"
                        },
                        "userName": {
                          "type": "string",
                          "example": "test"
                        },
                        "userPhoto": {
                          "type": "string",
                          "example": "http://example.com/photo.jpg"
                        },
                        "content": {
                          "type": "string",
                          "example": "這是一個測試貼文。"
                        },
                        "image": {
                          "type": "string",
                          "example": "http://nexample.com/image.jpg"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-02-28T21:44:04.703Z"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "搜尋失敗",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": false
                  },
                  "message": {
                    "type": "string",
                    "example": "搜尋異常"
                  },
                  "errMsg": {
                    "type": "string",
                    "example": "errMsg"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts - 新增貼文"
        ],
        "description": "新增貼文 API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "資料格式",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "六角"
                },
                "userPhoto": {
                  "type": "string",
                  "example": "http://example.com/photo.jpg"
                },
                "content": {
                  "type": "string",
                  "example": "這是一篇前端貼文"
                },
                "image": {
                  "type": "string",
                  "example": "http://nexample.com/image.jpg"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "新增成功",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": true
                  },
                  "message": {
                    "type": "string",
                    "example": "新增成功"
                  },
                  "post": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "662ec324ad9afd608b5c6d88"
                        },
                        "userName": {
                          "type": "string",
                          "example": "test"
                        },
                        "userPhoto": {
                          "type": "string",
                          "example": "http://example.com/photo.jpg"
                        },
                        "content": {
                          "type": "string",
                          "example": "這是一個測試貼文。"
                        },
                        "image": {
                          "type": "string",
                          "example": "http://nexample.com/image.jpg"
                        },
                        "createdAt": {
                          "type": "string",
                          "example": "2024-02-28T21:44:04.703Z"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "新增失敗",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "success": {
                    "type": "boolean",
                    "example": false
                  },
                  "message": {
                    "type": "string",
                    "example": "新增貼文失敗"
                  },
                  "errMsg": {
                    "type": "string",
                    "example": "Post validation failed: content: Content 未填寫, userName: 貼文姓名未填寫"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    }
  }
}